refactor controllers to a single package
Cleaner codebase.
