+++ src/main/java/sysc4806/project24/mini_survey_monkey/controllers/CreateController.java
@@ -1,4 +1,4 @@
-package sysc4806.project24.mini_survey_monkey;
+package sysc4806.project24.mini_survey_monkey.controllers;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
